Renamed exceptions.py as arc_exceptions.py
