+++ arc/mainTest.py
@@ -18,7 +18,7 @@ from rmgpy.molecule.molecule import Molecule
 from arc.main import ARC
 from arc.species.species import ARCSpecies
 from arc.settings import arc_path
-from arc.exceptions import InputError
+from arc.arc_exceptions import InputError
 
 ################################################################################
