+++ arc/processor.py
@@ -17,7 +17,7 @@ from rmgpy.species import Species
 import arc.rmgdb as rmgdb
 from arc.job.inputs import input_files
 from arc import plotter
-from arc.exceptions import SchedulerError, RotorError
+from arc.arc_exceptions import SchedulerError, RotorError
 from arc.species.species import determine_rotor_symmetry, determine_rotor_type
 
 ##################################################################
