+++ arc/ts/atst.py
@@ -4,7 +4,7 @@
 from __future__ import (absolute_import, division, print_function, unicode_literals)
 import os
 
-from arc.exceptions import TSError
+from arc.arc_exceptions import TSError
 from arc.settings import arc_path
