+++ arc/ts/run_autotst.py
@@ -20,7 +20,7 @@ import logging
 from autotst.reaction import AutoTST_Reaction
 
 from arc.parser import get_xyz_string
-from arc.exceptions import TSError
+from arc.arc_exceptions import TSError
 from arc.settings import arc_path
